Preserve State type in devtools
